Driver, Vehicle CRUD Updated

+++ RedisServices/DriverService.js
@@ -5,8 +5,9 @@ const strings = require('../Res/strings');
 var client = require('./dataContext').redisClient;
 var lredis = require('./redis-local');
 
-var PREFIX = strings.redis_prefix_driver;
-var DRIVER_ID = strings.redis_id_driver;
+const PREFIX = strings.redis_prefix_driver;
+const DRIVER_ID = strings.redis_id_driver;
+const DRIVER_LIST = strings.redis_prefix_driver_list;
 
 class DriverService {
   getDrivers() {
@@ -18,6 +19,20 @@ class DriverService {
         })).then(drivers => {
           resolve(drivers);
         })
+      })
+    });
+  }
+  getLocationDrivers(location) {
+    return new Promise((resolve, reject) => {
+      client.smembers(DRIVER_LIST + location, (err, ids) => {
+        if (err) resolve([]);
+        else {
+          Promise.all(ids.map(id => {
+            return lredis.hgetall(PREFIX + id);
+          })).then(drivers => {
+            resolve(drivers);
+          })
+        }
       })
     });
   }
@@ -35,6 +50,7 @@ class DriverService {
         if (err) resolve({ success: false, message: strings.string_response_error });
         driver.id = id;
         client.hmset(PREFIX + id, driver);
+        client.sadd(DRIVER_LIST + driver.location, id);
         resolve({ success: true, message: strings.string_response_added });
       });
     });
@@ -53,8 +69,15 @@ class DriverService {
   }
   removeDriver(id) {
     return new Promise((resolve, reject) => {
-      client.del(PREFIX + id);
-      resolve({ success: true, message: strings.string_response_removed });
+      this.getDriver(id).then(driver => {
+        if (driver.id == undefined) {
+          resolve({ success: false, message: strings.string_not_found_driver });
+        } else {
+          client.srem(DRIVER_LIST + driver.location, id);
+          client.del(PREFIX + id);
+          resolve({ success: true, message: strings.string_response_removed });
+        }
+      })
     })
   }
 }
@@ -65,5 +88,6 @@ class DriverService {
 // lastName:
 // mobile:
 // email:
+// location: 
 
 module.exports = DriverService;
